+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # china-regions changelog
 
+## 0.0.3
+
+- update `one-level-map.json`
+- update `two-levels-map.json`
+- update `three-levels-map.json`
+- update `three-levels-map-full.json`
+
 ## 0.0.2
 
 - add `one-level-map.json`